# Use SWAR technique for more efficient serialization