--- a/jsoniter-scala-core/jvm/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonWriter.scala
+++ b/jsoniter-scala-core/jvm/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonWriter.scala
@@ -3,10 +3,11 @@ package com.github.plokhotnyuk.jsoniter_scala.core
 import java.io.OutputStream
 import java.math.BigInteger
 import java.nio.charset.StandardCharsets
-import java.nio.{BufferOverflowException, ByteBuffer}
+import java.nio.{BufferOverflowException, ByteBuffer, ByteOrder}
 import java.time._
 import java.util.UUID
 import com.github.plokhotnyuk.jsoniter_scala.core.JsonWriter._
+import java.lang.invoke.{MethodHandles, VarHandle}
 import scala.annotation.tailrec
 import scala.{specialized => sp}
 
@@ -865,67 +866,35 @@ final class JsonWriter private[jsoniter_scala](
   }
 
   private[this] def writeUUID(mostSigBits: Long, leastSigBits: Long): Unit = count = {
-    val pos = ensureBufCapacity(38) // 38 == (new java.util.UUID(0, 0)).toString.length + 2
-    val buf = this.buf
     val ds = lowerCaseHexDigits
+    val vh = varHandleAsLong
+    val pos = ensureBufCapacity(40) // 38 == (new java.util.UUID(0, 0)).toString.length + 2
+    val buf = this.buf
     val mostSigBits1 = (mostSigBits >> 32).toInt
-    buf(pos) = '"'
-    val d1 = ds(mostSigBits1 >>> 24)
-    buf(pos + 1) = d1.toByte
-    buf(pos + 2) = (d1 >> 8).toByte
-    val d2 = ds((mostSigBits1 >> 16) & 0xFF)
-    buf(pos + 3) = d2.toByte
-    buf(pos + 4) = (d2 >> 8).toByte
-    val d3 = ds((mostSigBits1 >> 8) & 0xFF)
-    buf(pos + 5) = d3.toByte
-    buf(pos + 6) = (d3 >> 8).toByte
+    val d1 = ds(mostSigBits1 >>> 24) << 8
+    val d2 = ds((mostSigBits1 >> 16) & 0xFF).toLong << 24
+    val d3 = ds((mostSigBits1 >> 8) & 0xFF).toLong << 40
     val d4 = ds(mostSigBits1 & 0xFF)
-    buf(pos + 7) = d4.toByte
-    buf(pos + 8) = (d4 >> 8).toByte
+    vh.set(buf, pos, '"' | d1 | d2 | d3 | d4.toLong << 56)
     val mostSigBits2 = mostSigBits.toInt
-    buf(pos + 9) = '-'
-    val d5 = ds(mostSigBits2 >>> 24)
-    buf(pos + 10) = d5.toByte
-    buf(pos + 11) = (d5 >> 8).toByte
-    val d6 = ds((mostSigBits2 >> 16) & 0xFF)
-    buf(pos + 12) = d6.toByte
-    buf(pos + 13) = (d6 >> 8).toByte
-    buf(pos + 14) = '-'
+    val d5 = ds(mostSigBits2 >>> 24) << 16
+    val d6 = ds((mostSigBits2 >> 16) & 0xFF).toLong << 32
     val d7 = ds((mostSigBits2 >> 8) & 0xFF)
-    buf(pos + 15) = d7.toByte
-    buf(pos + 16) = (d7 >> 8).toByte
-    val d8 = ds(mostSigBits2 & 0xFF)
-    buf(pos + 17) = d8.toByte
-    buf(pos + 18) = (d8 >> 8).toByte
+    vh.set(buf, pos + 8, d4 >> 8 | d5 | d6 | d7.toLong << 56 | 0x2D000000002D00L) // 0x2D000000002D00L == '-'.toLong << 48 | '-' << 8
+    val d8 = ds(mostSigBits2 & 0xFF) << 8
     val leastSigBits1 = (leastSigBits >> 32).toInt
-    buf(pos + 19) = '-'
-    val d9 = ds(leastSigBits1 >>> 24)
-    buf(pos + 20) = d9.toByte
-    buf(pos + 21) = (d9 >> 8).toByte
-    val d10 = ds((leastSigBits1 >> 16) & 0xFF)
-    buf(pos + 22) = d10.toByte
-    buf(pos + 23) = (d10 >> 8).toByte
-    buf(pos + 24) = '-'
-    val d11 = ds((leastSigBits1 >> 8) & 0xFF)
-    buf(pos + 25) = d11.toByte
-    buf(pos + 26) = (d11 >> 8).toByte
-    val d12 = ds(leastSigBits1 & 0xFF)
-    buf(pos + 27) = d12.toByte
-    buf(pos + 28) = (d12 >> 8).toByte
+    val d9 = ds(leastSigBits1 >>> 24).toLong << 32
+    val d10 = ds((leastSigBits1 >> 16) & 0xFF).toLong << 48
+    vh.set(buf, pos + 16, d7 >> 8 | d8 | d9 | d10 | 0x2D000000) // 0x2D000000 == '-' << 24
+    val d11 = ds((leastSigBits1 >> 8) & 0xFF) << 8
+    val d12 = ds(leastSigBits1 & 0xFF).toLong << 24
     val leastSigBits2 = leastSigBits.toInt
-    val d13 = ds(leastSigBits2 >>> 24)
-    buf(pos + 29) = d13.toByte
-    buf(pos + 30) = (d13 >> 8).toByte
+    val d13 = ds(leastSigBits2 >>> 24).toLong << 40
     val d14 = ds((leastSigBits2 >> 16) & 0xFF)
-    buf(pos + 31) = d14.toByte
-    buf(pos + 32) = (d14 >> 8).toByte
-    val d15 = ds((leastSigBits2 >> 8) & 0xFF)
-    buf(pos + 33) = d15.toByte
-    buf(pos + 34) = (d15 >> 8).toByte
-    val d16 = ds(leastSigBits2 & 0xFF)
-    buf(pos + 35) = d16.toByte
-    buf(pos + 36) = (d16 >> 8).toByte
-    buf(pos + 37) = '"'
+    vh.set(buf, pos + 24, '-' | d11 | d12| d13 | d14.toLong << 56)
+    val d15 = ds((leastSigBits2 >> 8) & 0xFF) << 8
+    val d16 = ds(leastSigBits2 & 0xFF).toLong << 24
+    vh.set(buf, pos + 32, d14 >> 8 | d15 | d16 | 0x220000000000L) // 0x220000000000L == '"'.toLong << 40
     pos + 38
   }
 
@@ -1217,22 +1186,14 @@ final class JsonWriter private[jsoniter_scala](
   }
 
   private[this] def writeBoolean(x: Boolean): Unit = count = {
-    val pos = ensureBufCapacity(5) // false.toString.length
+    val vh = varHandleAsLong
+    val pos = ensureBufCapacity(8) // bytes in Long
     val buf = this.buf
-    if (x) {
-      buf(pos) = 't'
-      buf(pos + 1) = 'r'
-      buf(pos + 2) = 'u'
-      buf(pos + 3) = 'e'
-      pos + 4
-    } else {
-      buf(pos) = 'f'
-      buf(pos + 1) = 'a'
-      buf(pos + 2) = 'l'
-      buf(pos + 3) = 's'
-      buf(pos + 4) = 'e'
-      pos + 5
-    }
+    val v =
+      if (x) 0x65757274L
+      else 0x65736c6166L
+    vh.set(buf, pos, v)
+    pos + 4 + (v >> 38).toInt
   }
 
   private[this] def writeByte(x: Byte): Unit = count = {
@@ -1712,22 +1673,16 @@ final class JsonWriter private[jsoniter_scala](
   }
 
   private[this] def write8Digits(q0: Int, pos: Int, buf: Array[Byte], ds: Array[Short]): Int = {
+    val vh = varHandleAsLong
     val y1 = q0 * 140737489L // James Anhalt's algorithm for 8 digits: https://jk-jeon.github.io/posts/2022/02/jeaiii-algorithm/
-    val d1 = ds((y1 >>> 47).toInt)
-    buf(pos) = d1.toByte
-    buf(pos + 1) = (d1 >> 8).toByte
     val y2 = (y1 & 0x7FFFFFFFFFFFL) * 100
-    val d2 = ds((y2 >>> 47).toInt)
-    buf(pos + 2) = d2.toByte
-    buf(pos + 3) = (d2 >> 8).toByte
     val y3 = (y2 & 0x7FFFFFFFFFFFL) * 100
-    val d3 = ds((y3 >>> 47).toInt)
-    buf(pos + 4) = d3.toByte
-    buf(pos + 5) = (d3 >> 8).toByte
     val y4 = (y3 & 0x7FFFFFFFFFFFL) * 100
-    val d4 = ds((y4 >>> 47).toInt)
-    buf(pos + 6) = d4.toByte
-    buf(pos + 7) = (d4 >> 8).toByte
+    val d1 = ds((y1 >>> 47).toInt)
+    val d2 = ds((y2 >>> 47).toInt) << 16
+    val d3 = ds((y3 >>> 47).toInt).toLong << 32
+    val d4 = ds((y4 >>> 47).toInt).toLong << 48
+    vh.set(buf, pos, d1 | d2 | d3 | d4)
     pos + 8
   }
 
@@ -2170,6 +2125,8 @@ final class JsonWriter private[jsoniter_scala](
 }
 
 object JsonWriter {
+  private final val varHandleAsLong: VarHandle =
+    MethodHandles.byteArrayViewVarHandle(classOf[Array[Long]], ByteOrder.LITTLE_ENDIAN)
   private final val isGraalVM: Boolean =
     Option(System.getProperty("java.vendor.version")).getOrElse(System.getProperty("java.vm.name")).contains("GraalVM") ||
       java.lang.management.ManagementFactory.getRuntimeMXBean.getInputArguments.contains("-XX:+UseJVMCICompiler")